+++ watch4logs.py
@@ -10,9 +10,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# version 3.0.0 - 01/28/2025 - IP Geolocation integration, table & column renames, refinements - see changelog
+# version 3.2.0 - 02/01/2025 - MariaDB compatible and Log Rotation - see changelog
 #
-# Copyright 2024 Will Raymond <farmfreshsoftware@gmail.com>
+# Copyright 2024-2025 Will Raymond <farmfreshsoftware@gmail.com>
 #
 # CHANGELOG.md in repository - https://github.com/WillTheFarmer/apache-logs-to-mysql
 """
@@ -28,31 +28,20 @@ from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
 from dotenv import load_dotenv
 from logs2mysql import processLogs
+# Readability of event messages in console
+from color import fg
 load_dotenv() # Loads variables from .env into the environment
 watch_path = os.getenv('WATCH_PATH')
 watch_recursive = bool(int(os.getenv('WATCH_RECURSIVE')))
 watch_interval = int(os.getenv('WATCH_INTERVAL'))
 watch_log = int(os.getenv('WATCH_LOG'))
-# Readability of event messages in console
-class bcolors:
-    GREEN = '\33[32m'
-    GREENER = '\033[92m'
-    ERROR = '\33[41m' # CREDBG - red background
-    HEADER = '\033[95m'
-    REPEAT = '\033[94m'
-    ALERT = '\033[96m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-    ENDC = '\033[0m'
 class importLogs(FileSystemEventHandler):
   processFiles = 0
   def on_any_event(self, event):
     if event.event_type == 'created' and event.is_directory == False:
       setattr(importLogs, 'processFiles', 1)
       if watch_log >= 1:
-        print(bcolors.ALERT + "*files to process*" + bcolors.ENDC)
+        print(fg.ALERT, "*files to process*" , fg.RESET)
 
 if __name__ == "__main__":
   event_handler = importLogs()
@@ -62,7 +51,7 @@ if __name__ == "__main__":
   try:
     while True:
       if watch_log >= 2:
-        print(bcolors.REPEAT + "check for new files" + bcolors.ENDC)
+        print(fg.REPEAT, "check for new files" , fg.RESET)
       if importLogs.processFiles == 1:
         setattr(importLogs, 'processFiles', 0)
         processLogs()